CCTV STREAMING SERVER/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
opencv-python==4.8.1.78
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
python-dotenv==1.0.0